claude-code/multi-file-refactor/financial-automation/requirements.txt
PyMuPDF
pypdf
rapidocr-onnxruntime
PyYAML
